scratch-webjar-backend/src/main/java/scratch/webjar/backend/security/UserDetailsFactory.java
package scratch.webjar.backend.security;

import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.stereotype.Component;
import scratch.webjar.backend.domain.User;

import java.util.Collection;

import static java.util.Collections.emptySet;

/**
 * @author Karl Bennett
 */
@Component
public class UserDetailsFactory {

    public UserDetails create(User user) {
        return new UserDetails() {
            @Override
            public Collection<? extends GrantedAuthority> getAuthorities() {
                return emptySet();
            }

            @Override
            public String getPassword() {
                return user.getPassword();
            }

            @Override
            public String getUsername() {
                return user.getEmail();
            }

            @Override
            public boolean isAccountNonExpired() {
                return true;
            }

            @Override
            public boolean isAccountNonLocked() {
                return true;
            }

            @Override
            public boolean isCredentialsNonExpired() {
                return true;
            }

            @Override
            public boolean isEnabled() {
                return true;
            }
        };
    }
}
